splits away project_db utils
